+++ src/mmrelay/meshtastic_utils.py
@@ -144,6 +144,27 @@ from mmrelay.meshtastic.executors import (
     reset_executor_degraded_state,
     shutdown_shared_executors,
 )
+from mmrelay.meshtastic.health import (
+    _claim_health_probe_response_and_maybe_calibrate,
+    _extract_packet_request_id,
+    _failed_probe_ack_state_error,
+    _handle_probe_ack_callback,
+    _is_health_probe_response_packet,
+    _metadata_probe_ack_timeout_error,
+    _missing_ack_state_error,
+    _missing_local_node_ack_state_error,
+    _missing_probe_transport_error,
+    _missing_probe_wait_error,
+    _missing_received_nak_error,
+    _probe_device_connection,
+    _prune_health_probe_tracking,
+    _reset_probe_ack_state,
+    _seed_connect_time_skew,
+    _set_probe_ack_flag_from_packet,
+    _track_health_probe_request_id,
+    _wait_for_probe_ack,
+    requires_continuous_health_monitor,
+)
 from mmrelay.meshtastic.messaging import (
     _get_node_display_name,
     _get_packet_details,
@@ -520,482 +541,8 @@ def _coerce_bool(value: Any, default: bool, setting_name: str) -> bool:
 # for backward-compatible patch targets (tests patch mmrelay.meshtastic_utils.*).
 
 
-def _extract_packet_request_id(packet: Any) -> int | None:
-    """
-    Extract a request ID from a Meshtastic packet dict, if present.
-    """
-    if not isinstance(packet, dict):
-        return None
-
-    candidates: list[Any] = [packet.get("requestId"), packet.get("request_id")]
-    decoded = packet.get("decoded")
-    if isinstance(decoded, dict):
-        candidates.extend(
-            [
-                decoded.get("requestId"),
-                decoded.get("request_id"),
-            ]
-        )
-
-    for candidate in candidates:
-        parsed = _coerce_positive_int_id(candidate)
-        if parsed is not None:
-            return parsed
-    return None
-
-
-def _prune_health_probe_tracking(now: float | None = None) -> None:
-    """
-    Remove expired health-probe request IDs from the in-memory tracking map.
-    """
-    current = now if now is not None else time.monotonic()
-    expired_ids = [
-        request_id
-        for request_id, deadline in _health_probe_request_deadlines.items()
-        if deadline <= current
-    ]
-    for request_id in expired_ids:
-        _health_probe_request_deadlines.pop(request_id, None)
-
-
-def _track_health_probe_request_id(
-    raw_request_id: Any, timeout_secs: float
-) -> int | None:
-    """
-    Track a newly sent health-probe request ID for response log classification.
-
-    Returns the normalized request ID if tracking succeeded.
-    """
-    request_id = _coerce_positive_int_id(raw_request_id)
-    if request_id is None:
-        return None
-
-    expires_at = (
-        time.monotonic() + max(float(timeout_secs), 1.0) + HEALTH_PROBE_TRACK_GRACE_SECS
-    )
-    with _health_probe_request_lock:
-        _prune_health_probe_tracking()
-        _health_probe_request_deadlines[request_id] = expires_at
-    return request_id
-
-
-def _seed_connect_time_skew(rx_time: float) -> bool:
-    """Seed rxTime clock skew from an early packet if not yet calibrated.
-
-    Returns:
-        bool: True when a new skew value was calibrated for this packet.
-    """
-    global _relay_rx_time_clock_skew_secs
-    global _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs
-
-    if rx_time <= 0:
-        return False
-
-    calibrated_skew: float = 0.0
-    startup_age: float = 0.0
-    calibrated_from_reconnect_prestart: bool = False
-
-    now_wall = time.time()
-    now_monotonic = time.monotonic()
-    observed_skew = now_wall - rx_time
-
-    with _relay_rx_time_clock_skew_lock:
-        if _relay_rx_time_clock_skew_secs is not None:
-            return False
-
-        relay_start_time = RELAY_START_TIME
-        startup_age = max(0.0, now_monotonic - _relay_connection_started_monotonic_secs)
-        within_startup_window = startup_age <= RX_TIME_SKEW_BOOTSTRAP_WINDOW_SECS
-        startup_drain_active = _relay_startup_drain_deadline_monotonic_secs is not None
-        reconnect_bootstrap_deadline = (
-            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs
-        )
-        reconnect_bootstrap_active = (
-            reconnect_bootstrap_deadline is not None
-            and reconnect_bootstrap_deadline >= now_monotonic
-        )
-        if (
-            reconnect_bootstrap_deadline is not None
-            and reconnect_bootstrap_deadline < now_monotonic
-        ):
-            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs = None
-        packet_is_post_start = rx_time >= relay_start_time
-
-        if not packet_is_post_start and (
-            not within_startup_window
-            or (not startup_drain_active and not reconnect_bootstrap_active)
-        ):
-            return False
-
-        if abs(observed_skew) > RX_TIME_SKEW_BOOTSTRAP_MAX_SKEW_SECS:
-            logger.debug(
-                "Skipping rxTime skew bootstrap %.3f seconds outside startup limit %.3f",
-                observed_skew,
-                RX_TIME_SKEW_BOOTSTRAP_MAX_SKEW_SECS,
-            )
-            return False
-
-        _relay_rx_time_clock_skew_secs = observed_skew
-        calibrated_skew = observed_skew
-        calibrated_from_reconnect_prestart = (
-            not packet_is_post_start
-            and not startup_drain_active
-            and reconnect_bootstrap_active
-        )
-        if calibrated_from_reconnect_prestart:
-            # Consume the one-shot reconnect bootstrap allowance.
-            _relay_reconnect_prestart_bootstrap_deadline_monotonic_secs = None
-
-    if packet_is_post_start:
-        logger.debug(
-            "Calibrated rxTime clock skew from connect-time packet to %.3f seconds",
-            calibrated_skew,
-        )
-    elif calibrated_from_reconnect_prestart:
-        logger.debug(
-            "Bootstrapped rxTime clock skew from reconnect packet to %.3f seconds (startup_age=%.3f seconds)",
-            calibrated_skew,
-            startup_age,
-        )
-    else:
-        logger.debug(
-            "Bootstrapped rxTime clock skew from startup packet to %.3f seconds (startup_age=%.3f seconds)",
-            calibrated_skew,
-            startup_age,
-        )
-
-    return True
-
-
-def _is_health_probe_response_packet(packet: dict[str, Any], interface: Any) -> bool:
-    """
-    Determine if an inbound packet is a tracked health-probe response.
-    """
-    request_id = _extract_packet_request_id(packet)
-    if request_id is None:
-        return False
-
-    sender = _coerce_int_id(packet.get("from"))
-    local_num_raw = getattr(getattr(interface, "myInfo", None), "my_node_num", None)
-    if local_num_raw is None:
-        local_num_raw = getattr(getattr(interface, "localNode", None), "nodeNum", None)
-    local_num = _coerce_int_id(local_num_raw)
-    if sender is not None and local_num is not None and sender != local_num:
-        return False
-
-    with _health_probe_request_lock:
-        _prune_health_probe_tracking()
-        return request_id in _health_probe_request_deadlines
-
-
-def _claim_health_probe_response_and_maybe_calibrate(
-    packet: dict[str, Any], interface: Any, rx_time: float
-) -> bool:
-    """
-    Atomically claim a tracked health-probe response and calibrate skew once.
-
-    Lock order intentionally matches connect_meshtastic():
-    _health_probe_request_lock -> _relay_rx_time_clock_skew_lock.
-    """
-    global _relay_rx_time_clock_skew_secs
-
-    request_id = _extract_packet_request_id(packet)
-    if request_id is None:
-        return False
-
-    sender = _coerce_int_id(packet.get("from"))
-    local_num_raw = getattr(getattr(interface, "myInfo", None), "my_node_num", None)
-    if local_num_raw is None:
-        local_num_raw = getattr(getattr(interface, "localNode", None), "nodeNum", None)
-    local_num = _coerce_int_id(local_num_raw)
-    if sender is not None and local_num is not None and sender != local_num:
-        return False
-
-    observed_skew: float | None = None
-    calibrated_now = False
-    with _health_probe_request_lock:
-        _prune_health_probe_tracking()
-        if request_id not in _health_probe_request_deadlines:
-            return False
-
-        # Claim request ID so late duplicates cannot recalibrate.
-        _health_probe_request_deadlines.pop(request_id, None)
-
-        if rx_time > 0:
-            observed_skew = time.time() - rx_time
-            if abs(observed_skew) > RX_TIME_SKEW_BOOTSTRAP_MAX_SKEW_SECS:
-                logger.debug(
-                    "[HEALTH_CHECK] Skipping rxTime clock skew calibration %.3f seconds outside startup limit %.3f",
-                    observed_skew,
-                    RX_TIME_SKEW_BOOTSTRAP_MAX_SKEW_SECS,
-                )
-            else:
-                with _relay_rx_time_clock_skew_lock:
-                    if _relay_rx_time_clock_skew_secs is None:
-                        _relay_rx_time_clock_skew_secs = observed_skew
-                        calibrated_now = True
-
-    if calibrated_now and observed_skew is not None:
-        logger.debug(
-            "[HEALTH_CHECK] Calibrated rxTime clock skew to %.3f seconds",
-            observed_skew,
-        )
-    return True
-
-
-def _set_probe_ack_flag_from_packet(local_node: Any, packet: Any) -> bool:
-    """
-    Best-effort fallback for ACK packets missing routing metadata.
-
-    Some Meshtastic library versions can invoke ACK callbacks with packet shapes
-    that do not include `decoded.routing`, which causes `Node.onAckNak()` to
-    raise `KeyError("routing")`. For health probes we only need an ACK/NAK
-    signal, so this helper sets the same acknowledgment flags used by
-    `waitForAckNak()`.
-
-    Parameters:
-        local_node (Any): Meshtastic local node object expected to expose `iface`.
-        packet (Any): Callback packet payload (typically a dict).
-
-    Returns:
-        bool: True if a fallback ACK flag was set, False otherwise.
-    """
-    iface = getattr(local_node, "iface", None)
-    ack_state = getattr(iface, "_acknowledgment", None)
-    if ack_state is None:
-        return False
-
-    sender_raw = packet.get("from") if isinstance(packet, dict) else None
-    local_num = getattr(getattr(iface, "localNode", None), "nodeNum", None)
-
-    sender_num = _coerce_int_id(sender_raw)
-
-    if (
-        sender_num is not None
-        and local_num is not None
-        and sender_num == local_num
-        and hasattr(ack_state, "receivedImplAck")
-    ):
-        ack_state.receivedImplAck = True
-        return True
-
-    if hasattr(ack_state, "receivedAck"):
-        ack_state.receivedAck = True
-        return True
-
-    return False
-
-
-def _missing_local_node_ack_state_error() -> RuntimeError:
-    """
-    Build the error raised when local node ACK state is unavailable.
-    """
-    return RuntimeError("Meshtastic local node missing acknowledgment state")
-
-
-def _missing_received_nak_error() -> RuntimeError:
-    """
-    Build the error raised when ACK state cannot represent NAK responses.
-    """
-    return RuntimeError("Meshtastic acknowledgment state missing receivedNak")
-
-
-def _failed_probe_ack_state_error() -> RuntimeError:
-    """
-    Build the error raised when a probe response cannot set ACK/NAK state.
-    """
-    return RuntimeError("Failed to set ACK state from health probe response")
-
-
-def _missing_ack_state_error() -> RuntimeError:
-    """
-    Build the error raised when client ACK state is unavailable.
-    """
-    return RuntimeError("Meshtastic client missing acknowledgment state")
-
-
-def _metadata_probe_ack_timeout_error(timeout_secs: float) -> TimeoutError:
-    """
-    Build the timeout error raised when metadata probe ACK wait exceeds limit.
-    """
-    return TimeoutError(
-        f"Timed out waiting for metadata probe ACK after {timeout_secs:.1f} seconds"
-    )
-
-
-def _missing_probe_transport_error() -> RuntimeError:
-    """
-    Build the error raised when client cannot send metadata probe packets.
-    """
-    return RuntimeError("Meshtastic client cannot perform metadata liveness probe")
-
-
-def _missing_probe_wait_error() -> RuntimeError:
-    """
-    Build the error raised when client cannot wait for metadata probe ACKs.
-    """
-    return RuntimeError("Meshtastic client cannot wait for metadata probe ACK")
-
-
-def _reset_probe_ack_state(ack_state: Any) -> None:
-    """
-    Reset health-probe acknowledgment flags on a Meshtastic ACK state object.
-
-    Uses the object's `reset()` method when available; otherwise manually clears
-    known ACK/NAK flags for compatibility with test doubles and older interfaces.
-    """
-    reset = getattr(ack_state, "reset", None)
-    if callable(reset):
-        reset()
-        return
-
-    for attr in ("receivedAck", "receivedNak", "receivedImplAck"):
-        if hasattr(ack_state, attr):
-            setattr(ack_state, attr, False)
-
-
-def _handle_probe_ack_callback(local_node: Any, packet: Any) -> None:
-    """
-    Handle health-probe response packets across routing/admin response shapes.
-
-    For `get_device_metadata_request`, Meshtastic responses may be:
-    - A ROUTING_APP ACK/NAK packet with `decoded.routing.errorReason`, or
-    - An ADMIN_APP response packet that includes `requestId` but no
-      `decoded.routing`.
-
-    We treat either as forward progress for liveness probing by setting the same
-    acknowledgment flags used by `waitForAckNak()`.
-    """
-    iface = getattr(local_node, "iface", None)
-    ack_state = getattr(iface, "_acknowledgment", None)
-    if ack_state is None:
-        raise _missing_local_node_ack_state_error()
-
-    decoded = packet.get("decoded") if isinstance(packet, dict) else None
-    routing = decoded.get("routing") if isinstance(decoded, dict) else None
-    if isinstance(routing, dict):
-        error_reason = routing.get("errorReason")
-        if error_reason and error_reason != "NONE":
-            if hasattr(ack_state, "receivedNak"):
-                ack_state.receivedNak = True
-                return
-            raise _missing_received_nak_error()
-
-    if _set_probe_ack_flag_from_packet(local_node, packet):
-        return
-
-    raise _failed_probe_ack_state_error()
-
-
-def _wait_for_probe_ack(client: Any, timeout_secs: float) -> None:
-    """
-    Wait for ACK/NAK flags with a bounded timeout for health probes.
-
-    Uses the interface acknowledgment object directly so probe duration is
-    capped independently of the interface-wide timeout setting.
-    """
-    ack_state = getattr(client, "_acknowledgment", None)
-    if ack_state is None:
-        raise _missing_ack_state_error()
-
-    ack_attrs = ("receivedAck", "receivedNak", "receivedImplAck")
-
-    deadline = time.monotonic() + timeout_secs
-    while time.monotonic() < deadline:
-        if any(bool(getattr(ack_state, attr, False)) for attr in ack_attrs):
-            _reset_probe_ack_state(ack_state)
-            return
-        remaining = deadline - time.monotonic()
-        if remaining <= 0:
-            break
-        time.sleep(min(ACK_POLL_INTERVAL_SECS, remaining))
-
-    # Final check catches ACK/NAK updates that may land near the deadline.
-    if any(bool(getattr(ack_state, attr, False)) for attr in ack_attrs):
-        _reset_probe_ack_state(ack_state)
-        return
-
-    raise _metadata_probe_ack_timeout_error(timeout_secs)
-
-
-def _probe_device_connection(
-    client: Any, timeout_secs: float = DEFAULT_MESHTASTIC_OPERATION_TIMEOUT
-) -> None:
-    """
-    Send a metadata admin request and wait for an acknowledgment.
-
-    This uses the public sendData API instead of the private _sendAdmin
-    to ensure compatibility across Serial, TCP, and BLE interfaces.
-
-    Note: The onResponse callback must handle both routing ACK packets and
-    admin response packets for this request shape. Some Meshtastic versions
-    provide callback payloads without `decoded.routing`; this function uses a
-    guarded callback plus a bounded ACK wait to avoid long probe stalls.
-
-    Parameters:
-        client (Any): Meshtastic client/interface instance used for probe send and
-            ACK tracking.
-        timeout_secs (float): Maximum seconds to wait for probe ACK/NAK before
-            raising TimeoutError.
-    """
-    local_node = getattr(client, "localNode", None)
-    if local_node is None or not callable(getattr(client, "sendData", None)):
-        raise _missing_probe_transport_error()
-
-    # Clear stale ACK/NAK flags so this probe cannot "pass" on prior traffic.
-    ack_state = getattr(client, "_acknowledgment", None)
-    if ack_state is None:
-        ack_state = getattr(getattr(local_node, "iface", None), "_acknowledgment", None)
-    if ack_state is not None:
-        _reset_probe_ack_state(ack_state)
-
-    request = admin_pb2.AdminMessage()
-    request.get_device_metadata_request = True
-    # Use the public sendData API instead of private _sendAdmin
-    node_num = getattr(local_node, "nodeNum", None)
-    destination_id = node_num if node_num is not None else "^local"
-    sent_packet = client.sendData(
-        request.SerializeToString(),
-        destinationId=destination_id,
-        portNum=portnums_pb2.PortNum.ADMIN_APP,
-        wantAck=True,
-        wantResponse=True,
-        onResponse=functools.partial(_handle_probe_ack_callback, local_node),
-    )
-    request_id = _track_health_probe_request_id(
-        (
-            getattr(sent_packet, "id", None)
-            if not isinstance(sent_packet, dict)
-            else sent_packet.get("id")
-        ),
-        timeout_secs,
-    )
-    if request_id is not None:
-        logger.debug(
-            "[HEALTH_CHECK] Sent metadata probe requestId=%s timeout=%.1fs",
-            request_id,
-            timeout_secs,
-        )
-    else:
-        logger.debug(
-            "[HEALTH_CHECK] Sent metadata probe timeout=%.1fs",
-            timeout_secs,
-        )
-
-    if getattr(client, "_acknowledgment", None) is not None:
-        _wait_for_probe_ack(client, timeout_secs)
-        return
-
-    if callable(getattr(client, "waitForAckNak", None)):
-        _run_blocking_with_timeout(
-            cast(Callable[[], Any], client.waitForAckNak),
-            timeout=timeout_secs,
-            label="metadata-probe-waitForAckNak",
-            timeout_log_level=logging.DEBUG,
-        )
-        return
-
-    raise _missing_probe_wait_error()
+# Health probe / ACK tracking / rxTime skew calibration — implemented in meshtastic.health,
+# re-exported here for backward-compatible patch targets (tests patch mmrelay.meshtastic_utils.*).
 
 
 def _submit_coro(
@@ -4412,37 +3959,8 @@ def on_meshtastic_message(packet: dict[str, Any], interface: Any) -> None:
         )
 
 
-def requires_continuous_health_monitor(config: dict[str, Any]) -> bool:
-    """
-    Return True when periodic health monitoring should run for the given config.
-
-    Health monitoring is disabled for BLE connections (which have real-time disconnect
-    detection) and when health_check.enabled is explicitly set to False.
-
-    Note: This function only returns False for "clean" exit cases (BLE or explicitly
-    disabled). Malformed health_check configs return True so that check_connection
-    can log appropriate warnings before exiting.
-
-    Args:
-        config: The full configuration dictionary.
-
-    Returns:
-        True if health monitoring should run continuously, False otherwise.
-    """
-    meshtastic_config = config.get(CONFIG_SECTION_MESHTASTIC)
-    if not isinstance(meshtastic_config, dict):
-        return DEFAULT_HEALTH_CHECK_ENABLED
-    if meshtastic_config.get(CONFIG_KEY_CONNECTION_TYPE) == CONNECTION_TYPE_BLE:
-        return False
-    health_config = meshtastic_config.get("health_check")
-    if health_config is None:
-        return DEFAULT_HEALTH_CHECK_ENABLED
-    if not isinstance(health_config, dict):
-        return True
-    raw_enabled = health_config.get("enabled", DEFAULT_HEALTH_CHECK_ENABLED)
-    return _coerce_bool(
-        raw_enabled, DEFAULT_HEALTH_CHECK_ENABLED, "health_check.enabled"
-    )
+# Health monitoring config interpretation — implemented in meshtastic.health,
+# re-exported here for backward-compatible patch targets (tests patch mmrelay.meshtastic_utils.*).
 
 
 async def check_connection() -> None:
